Rename webxdc-types to webxdc
The -types postfix was redundant given they were in the /types directory
